build: clean up build scripts
